Move picks->TopPicks and fetch real data
