+++ src/components/TopPicks/TopPicks.stories.tsx
@@ -1,7 +1,17 @@
 import React from "react";
 import { TopPicks } from "./TopPicks";
+import { css } from "emotion";
 
-export default { component: TopPicks, title: "TopPicks" };
+export default { component: TopPicks, title: "Pick" };
 
-export const Default = () => <TopPicks />;
-Default.story = { name: "default" };
+export const Default = () => (
+  <div
+    className={css`
+      width: 100%;
+      max-width: 600px;
+    `}
+  >
+    <TopPicks shortUrl="/p/39f5z" />
+  </div>
+);
+Default.story = { name: "Picks" };
